Improve typing, handle Numba failures, switch to pytest
This brings numba into a `use_numba` option, though it is still turned
off by default.  Numba is tested not to fail, and to provide values
that are reasonably close to non-numba calculations (tested randomly).
The sped-up version of bootstrap and jackknife is *only* used with
use_numba.

Overload typing is implemented for ci, and numerous minor type problem
have been improved.